supermarket-ai/requirements.txt
flask
requests
pandas
python-dotenv
twilio
sentence-transformers
faiss-cpu
gunicorn